chore:  Added more update commands

+++ SweetManagerIotWebService.API/Reservations/Domain/Model/Commands/Thermostat/UpdateThermostatTemperatureCommand.cs
@@ -2,4 +2,4 @@
 
 public record UpdateThermostatTemperatureCommand(
     int Id,
-    int? Temperature);
+    double? Temperature);
